118: Build AlbumDetail
