+++ src/UI/portal/src/pages/user/RolePage.tsx
@@ -1,51 +1,48 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { useSelector } from "react-redux";
-import { StoreState, useAppDispatch } from '../store';
-import { getAlbumsPagingAsyncThunk, getAlbumAlertMessagesAsyncThunk, getAllContentTypesAsyncThunk } from "../store/reducers/albumSlice";
-import AlbumPagingResponse from "../models/album/AlbumPagingResponse";
-import { ActionTypeGrid } from "../models/enums/ActionTypeGrid";
-import { useTranslation } from "react-i18next";
-import ModalCommon from "../components/shared/ModalCommon";
-import Pagination from "../components/shared/Pagination";
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import ModalCommon from '../../components/shared/ModalCommon';
+import { ActionTypeGrid } from '../../models/enums/ActionTypeGrid';
+import { useTranslation } from 'react-i18next';
+import { useDispatch, useSelector } from 'react-redux';
+import { StoreState } from '../../store';
+import Role from '../../models/role/Role';
+import Pagination from '../../components/shared/Pagination';
+import { getRoles } from '../../store/thunks/roleThunk';
 import { v4 as uuidv4 } from 'uuid';
-import dayjsCustom from "../utils/dayjs/dayjs-custom";
-import CreateAlbum from "../components/album/CreateAlbum";
-import UpdateAlbum from "../components/album/UpdateAlbum";
-import DeleteAlbum from "../components/album/DeleteAlbum";
+import CreateRole from '../../components/role/CreateRole';
+import DeleteRole from '../../components/role/DeleteRole';
+import UpdateRole from '../../components/role/UpdateRole';
 
-const AlbumPage: React.FC = () => {
+const RolePage: React.FC = () => {
     const [isOpen, setIsOpen] = useState(false);
     const [actionGrid, setActionGrid] = useState(ActionTypeGrid.CREATE);
 
     const [t] = useTranslation();
 
-    const albumState = useSelector((state: StoreState) => state.album);
-    const dispatch = useAppDispatch();
+    const roleState = useSelector((state: StoreState) => state.role);
+    const dispatch = useDispatch();
 
-    const albums = useMemo(() => albumState.albums, [albumState.albums]);
-    const [album, setAlbum] = useState<AlbumPagingResponse>(albums[0] || null);
+    const roles = useMemo(() => roleState.roles, [roleState.roles]);
+    const [role, setRole] = useState<Role>(roles[0] || null);
 
     // Paging
     const [pageIndex, setPageIndex] = useState(1);
     const [pageSize, setPageSize] = useState(5);
 
     useEffect(() => {
-        dispatch(getAlbumsPagingAsyncThunk({ pageNumber: pageIndex, pageSize }));
-        dispatch(getAllContentTypesAsyncThunk());
-        dispatch(getAlbumAlertMessagesAsyncThunk())
+        getRoles(pageIndex, pageSize)(dispatch);
     }, [dispatch, pageIndex, pageSize]);
 
-    const openModal = (actionGrid: ActionTypeGrid, album?: AlbumPagingResponse) => {
+    const openModal = (actionGrid: ActionTypeGrid, role?: Role) => {
         setActionGrid(actionGrid);
-        if (album) {
-            setAlbum(album);
+        if (role) {
+            setRole(role);
         }
         setIsOpen(true);
     }
 
     const closeModal = (isReload?: boolean) => {
         if (isReload) {
-            dispatch(getAlbumsPagingAsyncThunk({ pageNumber: pageIndex, pageSize }));
+            getRoles(pageIndex, pageSize)(dispatch);
         }
         setIsOpen(false);
     }
@@ -53,13 +50,14 @@ const AlbumPage: React.FC = () => {
     const BodyModal = useCallback((actionGrid: ActionTypeGrid) => {
         switch (actionGrid) {
             case ActionTypeGrid.CREATE:
-                return CreateAlbum;
+                return CreateRole;
             case ActionTypeGrid.EDIT:
-                return UpdateAlbum;
+                return UpdateRole;
             case ActionTypeGrid.DELETE:
-                return DeleteAlbum;
+                return DeleteRole;
+            default:
+                return React.Fragment;
         }
-
     }, []);
 
     return (
@@ -82,10 +80,10 @@ const AlbumPage: React.FC = () => {
                                                 <a href="crm-contacts.html#">CMS</a>
                                             </li>
                                             {/*end nav-item*/}
-                                            <li className="breadcrumb-item active">Albums</li>
+                                            <li className="breadcrumb-item active">Users</li>
                                         </ol>
                                     </div>
-                                    <h4 className="page-title">{t("album.title")}</h4>
+                                    <h4 className="page-title">{t("role.title")}</h4>
                                 </div>
                                 {/*end page-title-box*/}
                             </div>
@@ -98,7 +96,7 @@ const AlbumPage: React.FC = () => {
                                     <div className="card-header">
                                         <div className="row align-items-center">
                                             <div className="col">
-                                                <h4 className="card-title">{t('album.title_detail')}</h4>
+                                                <h4 className="card-title">{t('role.title_detail')}</h4>
                                             </div>
                                             {/*end col*/}
                                         </div>{" "}
@@ -110,31 +108,28 @@ const AlbumPage: React.FC = () => {
                                             <table className="table">
                                                 <thead>
                                                     <tr>
-                                                        <th>{t('album.id')}</th>
-                                                        <th>{t('album.album_title')}</th>
-                                                        <th>{t('album.album_description')}</th>
-                                                        <th>{t('album.album_content_types')}</th>
-                                                        <th>{t('album.created_on')}</th>
-                                                        <th>{t('album.updated_on')}</th>
-                                                        <th>{t('album.action')}</th>
+                                                        <th>{t('role.id')}</th>
+                                                        <th>{t('role.name')}</th>
+                                                        <th>{t('role.normalized_name')}</th>
+                                                        <th>{t('role.users')}</th>
+                                                        <th>{t('role.action')}</th>
                                                     </tr>
                                                 </thead>
                                                 <tbody>
-                                                    {albums.map((album) => (
+                                                    {roles.map((role) => (
                                                         <tr key={uuidv4()}>
-                                                            <td>{album.id}</td>
-                                                            <td>{album.title}</td>
-                                                            <td>{album.description}</td>
-                                                            <td>{album.contentTypes}</td>
-                                                            <td>{dayjsCustom.utc(album.createdOnUtc).local().format('DD-MM-YYYY HH:mm')}</td>
-                                                            <td>{album.updatedOnUtc && dayjsCustom.utc(album.updatedOnUtc).local().format('DD-MM-YYYY HH:mm')}</td>
+                                                            <td>{role.id}
+                                                            </td>
+                                                            <td>{role.name}</td>
+                                                            <td>{role.normalizedName}</td>
+                                                            <td>{role.users}</td>
                                                             <td>
                                                                 <button className="btn"
-                                                                    onClick={() => openModal(ActionTypeGrid.EDIT, album)}>
+                                                                    onClick={() => openModal(ActionTypeGrid.EDIT, role)}>
                                                                     <i className="fa-solid fa-pen text-secondary font-16"></i>
                                                                 </button>
                                                                 <button className="btn"
-                                                                    onClick={() => openModal(ActionTypeGrid.DELETE, album)}>
+                                                                    onClick={() => openModal(ActionTypeGrid.DELETE, role)}>
                                                                     <i className="fa-solid fa-trash text-secondary font-16"></i>
                                                                 </button>
                                                             </td>
@@ -147,7 +142,7 @@ const AlbumPage: React.FC = () => {
                                             <div className="col">
                                                 <button className="btn btn-outline-light btn-sm px-4"
                                                     onClick={() => openModal(ActionTypeGrid.CREATE)}>
-                                                    + {t('user.add_new')}
+                                                    + {t('role.add_new')}
                                                 </button>
                                             </div>
                                             <div className="col">
@@ -166,7 +161,7 @@ const AlbumPage: React.FC = () => {
                                                 <nav aria-label="...">
                                                     <Pagination
                                                         pageIndex={pageIndex}
-                                                        totalCounts={albumState.totalRecords}
+                                                        totalCounts={roleState.totalRecords}
                                                         pageSize={pageSize}
                                                         onPageChange={page => setPageIndex(page)} />
                                                     {/*end pagination*/}
@@ -190,11 +185,11 @@ const AlbumPage: React.FC = () => {
                 {/* end page content */}
             </div>
             <ModalCommon
-                props={{ modalIsOpen: isOpen, openModal, closeModal, album }}
+                props={{ modalIsOpen: isOpen, openModal, closeModal, role }}
                 Component={BodyModal(actionGrid)}
             />
         </>
     );
 };
 
-export default AlbumPage;
+export default RolePage;
